feat: Extract validation logic from Controller to Interceptors
