+++ src/main/java/nextstep/security/LoginAuthInterceptor.java
@@ -1,0 +1,37 @@
+package nextstep.security;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
+import nextstep.app.domain.Member;
+import nextstep.app.domain.MemberRepository;
+import org.springframework.stereotype.Component;
+import org.springframework.web.servlet.HandlerInterceptor;
+
+import java.util.Map;
+
+@Component
+public class LoginAuthInterceptor implements HandlerInterceptor {
+
+    private final MemberRepository memberRepository;
+
+    public static final String SPRING_SECURITY_CONTEXT_KEY = "SPRING_SECURITY_CONTEXT";
+
+    public LoginAuthInterceptor(MemberRepository memberRepository) {
+        this.memberRepository = memberRepository;
+    }
+
+    @Override
+    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
+        Map<String, String[]> parameterMap = request.getParameterMap();
+        String username = parameterMap.get("username")[0];
+        String password = parameterMap.get("password")[0];
+
+        Member member = memberRepository.findByEmail(username)
+                .filter(it -> it.matchPassword(password))
+                .orElseThrow(AuthenticationException::new);
+        HttpSession session = request.getSession();
+        session.setAttribute(SPRING_SECURITY_CONTEXT_KEY, member.toString());
+        return false;
+    }
+}
